requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
better-profanity==0.7.0
cachetools==4.2.2
certifi==2023.7.22
dnspython==2.4.1
joblib==1.3.2
numpy==1.24.4
pymongo==4.4.1
python-telegram-bot==13.15
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.2.0
tornado==6.1
tzlocal==5.2
